[PWX-25354] Setting max storage node per zone in Portworx
