+++ pkg/cloudprovider/cloudprovider.go
@@ -1,10 +1,13 @@
 package cloudprovider
 
 import (
+	"context"
 	"fmt"
 	"sync"
 
+	"github.com/sirupsen/logrus"
 	v1 "k8s.io/api/core/v1"
+	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/libopenstorage/cloudops"
 	"github.com/libopenstorage/operator/pkg/preflight"
@@ -59,6 +62,38 @@ func (d *defaultProvider) GetZone(node *v1.Node) (string, error) {
 // Get returns the cloud provider
 func Get() Ops {
 	return New(preflight.Instance().ProviderName())
+}
+
+// GetZoneMap returns zone map
+func GetZoneMap(k8sClient client.Client, filterLabelKey string, filterLabelValue string) (map[string]uint64, error) {
+	zoneMap := make(map[string]uint64)
+	nodeList := &v1.NodeList{}
+	err := k8sClient.List(context.TODO(), nodeList, &client.ListOptions{})
+	if err != nil {
+		return zoneMap, err
+	}
+	for _, node := range nodeList.Items {
+		if zone, err := Get().GetZone(&node); err == nil {
+			if len(filterLabelKey) > 0 {
+				value, ok := node.Labels[filterLabelKey]
+				// If provided filterLabelKey is not found or the value does not match with
+				// the provided filterLabelValue, no need to count in the zoneMap
+				if !ok || value != filterLabelValue {
+					if _, ok := zoneMap[zone]; !ok {
+						zoneMap[zone] = 0
+					}
+					continue
+				}
+				// provided filterLabel's value and key matched, let's count them in the zoneMap
+			}
+			instancesCount := zoneMap[zone]
+			zoneMap[zone] = instancesCount + 1
+		} else {
+			logrus.Errorf("count not find zone information: %v", err)
+			return nil, err
+		}
+	}
+	return zoneMap, nil
 }
 
 // New returns a new implementation of the cloud provider
